Move config.py to config dir
